--- a/Libraries/LLM/KVCache.swift
+++ b/Libraries/LLM/KVCache.swift
@@ -1,0 +1,104 @@
+// Copyright © 2024 Apple Inc.
+
+import Foundation
+import MLX
+
+/// Interface for Key/Value cache for LLMs.
+///
+/// See ``LLMModel/newCache(parameters:)-47tyu``
+public protocol KVCache: Evaluatable {
+
+    /// get the current offset
+    var offset: Int { get }
+
+    func update(keys: MLXArray, values: MLXArray) -> (MLXArray, MLXArray)
+}
+
+func createAdditiveCausalMask(n: Int, offset: Int) -> MLXArray {
+    let rinds = MLXArray(Int32(0) ..< Int32(offset + n))
+    let linds = offset != 0 ? MLXArray(Int32(offset) ..< Int32(offset + n)) : rinds
+    let mask = linds[0..., .newAxis] .< rinds[.newAxis]
+    return mask * Float32(-1e9)
+}
+
+/// create an attention mask using the parameters from the KVCache.
+///
+/// See also ``MultiHeadAttention/createAdditiveCausalMask(_:dtype:)`` -- same idea
+/// but doesn't honor the cache offset.
+public func createAttentionMask(h: MLXArray, cache: [KVCache]?) -> MLXArray? {
+    let t = h.dim(1)
+    if t > 1 {
+        var offset = 0
+        if let c = cache?.first {
+            offset = c.offset
+        }
+        return createAdditiveCausalMask(n: t, offset: offset)
+            .asType(h.dtype)
+    }
+    return nil
+}
+
+/// See https://github.com/ml-explore/mlx-examples/blob/main/llms/mlx_lm/models/base.py#L11
+class KVCacheSimple: KVCache, Evaluatable {
+    let kHeadDim: Int
+    let vHeadDim: Int
+    let kvHeads: Int
+
+    var keys: MLXArray?
+    var values: MLXArray?
+
+    var offset = 0
+    var step = 256
+
+    init(headDim: IntOrPair, kvHeads: Int) {
+        self.kHeadDim = headDim.first
+        self.vHeadDim = headDim.second
+        self.kvHeads = kvHeads
+    }
+
+    public func innerState() -> [MLXArray] {
+        [self.keys, self.values].compactMap { $0 }
+    }
+
+    func update(keys: MLXArray, values: MLXArray) -> (MLXArray, MLXArray) {
+        let previous = self.offset
+
+        let reset =
+            if let currentKeys = self.keys, (previous + keys.dim(2)) > currentKeys.dim(2) {
+                true
+            } else {
+                self.keys == nil
+            }
+        if reset {
+            let B = keys.dim(0)
+            let nSteps = (step + keys.dim(2) - 1) / step
+            let kShape = [B, kvHeads, nSteps * step, kHeadDim]
+            let vShape = [B, kvHeads, nSteps * step, vHeadDim]
+            let newK = MLXArray.zeros(kShape, dtype: keys.dtype)
+            let newV = MLXArray.zeros(vShape, dtype: values.dtype)
+
+            if var currentKeys = self.keys, var currentValues = self.values {
+                if previous % step != 0 {
+                    currentKeys = currentKeys[.ellipsis, ..<previous, 0...]
+                    currentValues = currentValues[.ellipsis, ..<previous, 0...]
+                }
+                self.keys = concatenated([currentKeys, newK], axis: 2)
+                self.values = concatenated([currentValues, newV], axis: 2)
+            } else {
+                self.keys = newK
+                self.values = newV
+            }
+        }
+
+        self.offset += keys.dim(2)
+
+        self.keys?[.ellipsis, previous ..< self.offset, 0...] = keys
+        self.values?[.ellipsis, previous ..< self.offset, 0...] = values
+
+        return (
+            self.keys![.ellipsis, ..<self.offset, 0...],
+            self.values![.ellipsis, ..<self.offset, 0...]
+        )
+    }
+
+}