… Libraries/LLM/LLMModel.swift
… Libraries/LLM/Llama.swift
… Libraries/LLM/Lora.swift
… Libraries/LLM/OpenELM.swift
… Libraries/LLM/Phi.swift
… Libraries/LLM/Phi3.swift
… Libraries/LLM/Qwen2.swift
… Libraries/LLM/Starcoder2.swift
… Libraries/LLM/Tokenizer.swift
… Package.swift
… Tools/llm-tool/LLMTool.swift
… mlx-swift-examples.xcodeproj/project.pbxproj
… mlx-swift-examples.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved