- sampling code compiled
- KVCache
- async_eval
- NaiveStreamingDetokenizer

My starting point with mlx v0.16.0:

```
./mlx-run llm-tool --model mlx-community/Phi-3-mini-4k-instruct-4bit -m 100000 --temperature 0 -p @/tmp/p.txt --cache-size 500 --temperature 0
```

Prompt is:

```
<|system|>
You are a helpful assistant.<|end|>
<|user|>
How to explain Internet for a medieval knight?<|end|>
<|assistant|>
```

Here are the initial performance numbers:

```
python Generation: 673 tokens, 114.404 tokens-per-sec, 5.873911874999999s
swift Generation: 673 tokens, 83.136954 tokens/s
```

Looking at the profiler we can see:

| Task | Python | Swift |
| --- | --- | --- |
| Wall Time | 9.6 | 12.1 |
| CPU Time | 10.94 | 9.91 |
| Metal Submit | 1.75 | 3.83 |
| Metal didComplete | 0.48 | 0.83 |
| mlx::core:scheduler | 2.92 | 3.63 |
| blas_thread_server | 2.9 | - |
| main | 1.67 | 0.74 |

I am not sure what the `blas_thread_server` is, but it is soaking up a bunch of CPU time — it looks like it is running async to the main computation, so other than efficiency isn’t affecting performance.  If we subtract the blas_thread_server from the CPU time it is clear the python code is doing less work.

I found the following:

- The sampling code in python is compiled — this made a pretty big difference
- The KVCache also made a big difference
- The async_eval in the python code helped a little
- The NaiveStreamingDetokenizer helped a little

The prompt splitting didn’t help here as the default split size is 512 and that is larger than the prompt.  The output in this case is only 673 tokens, so I don’t think that is particularly long, so some of these may not be getting the full benefit.

With these in place the performance is getting closer:

```
python Generation: 673 tokens, 114.404 tokens-per-sec, 5.873911874999999s
swift 673 tokens, 97.718867 tokens/s, 6.887104s
```

| Task | Python | Swift |
| --- | --- | --- |
| Wall Time | 9.6 | 9.4 |
| CPU Time | 10.94 | 7.6 |
| Metal Submit | 1.75 | 1.21 |
| Metal didComplete | 0.48 | 0.28 |
| mlx::core:scheduler | 2.92 | 1.99 |
| blas_thread_server | 2.9 | - |
| main | 1.67 | 0.69 |

We have a big win in CPU time and even total wall time (startup time I presume) but are still nearly a second slower on the generation, so there is still some more work needed.

# Part 2!

Here is the output from swift -- for comparison I had 114 tokens/s from python (though it was usually lower than that and I found the number to be noisy). 

```
Prompt:     21 tokens, 212.221945 tokens/s
Generation: 673 tokens, 114.662957 tokens/s, 5.869376s
```

Here is output from Instruments showing the synchronous evaluation (via `item()`), the `async_eval()` calls and the GPU activity in the python program:

<img width="1221" alt="image" src="https://github.com/user-attachments/assets/41eaccce-6591-4a58-8db7-cc6a3ce96a49">

Notice that the GPU stays active most of the time.

Here is a similar view of the swift program (at around 97 tokens/s) from part 1:

<img width="1222" alt="image" src="https://github.com/user-attachments/assets/0add6f4a-0dbe-4bbf-8432-72d29e62974a">

The pattern is similar, but there are ~1.5ms gaps where the GPU goes idle -- the scheduler didn't have any work to do.

The final version looks like this:

<img width="1214" alt="image" src="https://github.com/user-attachments/assets/19d0da35-2a09-4409-b780-b4e08fa30960">

The GPU is kept busy and we are hitting our performance targets!

The final changes resolved an issue with the `asyncEval()` and `item()`/`eval()` calls that were causing `item()` to wait for the completion of the `asyncEval()` -- see https://github.com/ml-explore/mlx-swift/pull/127.