@awni 2024-08-24
> With these in place the performance is getting closer:

> python Generation: 673 tokens, 114.404 tokens-per-sec, 5.873911874999999s
> swift 673 tokens, 97.718867 tokens/s, 6.887104s

That is very strange.. all of the fixes you pulled in should have gotten us to basically parity with Python but it's still substantially slower 🤔 

Could be some perf improvements in the updated MLX (17.0) as well, not sure what you were running.