@awni 2024-08-24
> I am not sure what the blas_thread_server is, but it is soaking up a bunch of CPU time 

That's not an MLX thing.. it looks like it's coming from [openblas](https://github.com/OpenMathLib/OpenBLAS/blob/develop/driver/others/blas_server.c#L379) which is probably from NumPy if I had to guess. Very odd though .. its possibly dependent on the numpy version / back-end