@davidkoski 2024-08-24
> @davidkoski any reason this is still in draft? It looks more or less ready for review/merge?

Still needs:

- some documentation
- another KVCache implementation or at least setting it up as a protocol so it can be used that way

I wanted to figure out the rest of the performance gap but I think this is useful enough that we could take it.

> Could be some perf improvements in the updated MLX (17.0) as well, not sure what you were running.

I was running 16.0 for both.

One thing I will check is the implementation of the LLM model code -- maybe that has diverged.