@davidkoski 2024-08-27
@awni this is ready for review.  It requires the changes from https://github.com/ml-explore/mlx-swift/pull/127 so I will have one more commit after I merge and tag that one.